Refactor estructura de datos para eventos deportivos y integración con API Sports
- Reestructuración de DTOs: movimiento de clases Status de EventsByLeagueResponse a Fixture
- Creación de nueva clase OddsResponse para manejo de momios
- Actualización de entidades: EventoDeportivo, Momio, Estado, Score, Goles y Valor
- Mejoras en repositorio EventoDeportivoRepository para nuevas consultas
- Expansión del servicio ApiSportService con nuevos métodos para procesamiento de datos
- Actualización de tareas programadas en ApiSportsScheduledTasks
- Refactoring general para mejor organización del código y compatibilidad con API externa

+++ backend/src/main/java/com/_bet/entity/eventoEntity/Valor.java
@@ -1,6 +1,5 @@
 package com._bet.entity.eventoEntity;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Entity;
@@ -21,7 +20,7 @@ public class Valor {
     private Long id;
 
     private String valor;
-    private BigDecimal odd;
+    private Double odd;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "momio_id", nullable = false)
